Enable the shadow compiler warning for GCC and clang
At some point in the past, the fmt library used to compile with
no-shadow so produced errors for us. Looks like they don't do that
anymore so we can get warning free compilation with this flag.

+++ include/nuclear-data-reader/nubase_data.hpp
@@ -258,7 +258,7 @@ namespace NUBASE
           full_data.substr(position.START_HALFLIFEERROR, (position.END_HALFLIFEERROR - position.START_HALFLIFEERROR));
       std::replace(hle.begin(), hle.end(), '>', ' ');
       std::replace(hle.begin(), hle.end(), '<', ' ');
-      return Converter::StringToNum<double>(hle, 0, hle.size());
+      return Converter::StringToNum<double>(hle, 0, static_cast<uint8_t>(hle.size()));
     }
 
     /**
